refactor(renter): improve host scanning and autopilot testing
* Change TestAutoPilotConfig to accept GougingSettings instead of AutopilotConfig
* Add manageScan and waitForScanCompletion methods for better scan control
* Update price tracking to use uint64 for host counts
* Improve price calculations to include TB unit conversions
* Add timeout and interval constants for scan management
* Enhance logging with more detailed price information
* Split host processing logic into separate batching function

+++ service/internal/renter/price_tracking.go
@@ -11,7 +11,7 @@ type PriceTracking struct {
 	maxStoragePrice types.Currency
 	maxUploadPrice  types.Currency
 	downloadPrices  []types.Currency
-	validHostsCount int
+	validHostsCount uint64
 	baseSettings    api.GougingSettings // Store the base settings
 }
 
@@ -47,7 +47,7 @@ func (pt *PriceTracking) UpdatePrices(hosts []Host) {
 		pt.downloadPrices = append(pt.downloadPrices, downloadTotal)
 	}
 
-	pt.validHostsCount += len(hosts)
+	pt.validHostsCount += uint64(len(hosts))
 }
 
 func (pt *PriceTracking) ComputeFinalPrices() api.GougingSettings {
